+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog ##
 #### dev ####
 * Fix patch error with ByRef while mixing primitive and wrapper types
+* Fix crash from SpireField of array type
 
 #### v3.15.0 ####
 * Allow mods to sideload other mods, loading them even if not selected